Pull Printer into separate file
